rename ValueDisplay to ValueInfoDisplay as it only shows the Parameter value info containers
