+++ frontend/src/modules/displays/ParameterValueDisplay.tsx
